Modified Java file

+++ src/CreditPaymentService.java
@@ -1,7 +1,7 @@
 public class CreditPaymentService {
-    public int calculate(double creditAmount, double annualInterestRate, int creditTime) {
+    public int calculate(int creditAmount, double annualInterestRate, int creditTime) {
         double monthlyInterestRate = annualInterestRate / 100 / 12;
-        double monthlyPayment = creditAmount * (monthlyInterestRate + monthlyInterestRate / (Math.pow(1 + monthlyInterestRate, creditTime) - 1));
-        return (int) monthlyPayment;
+        int monthlyPayment = (int) (creditAmount * (monthlyInterestRate + monthlyInterestRate / (Math.pow(1 + monthlyInterestRate, creditTime) - 1)));
+        return monthlyPayment;
     }
 }
